+++ c/basics.h
@@ -26,7 +26,7 @@ void put_be16(uint8_t *out, uint16_t i);
   find smallest index i in [0, sz) at which f(i) is true, assuming
   that f is ascending. Return sz if f(i) is false for all indices.
 */
-int binsearch(int sz, int (*f)(int k, void *args), void *args);
+int binsearch(size_t sz, int (*f)(size_t k, void *args), void *args);
 
 /*
   Frees a NULL terminated array of malloced strings. The array itself is also
