+++ c/slice.h
@@ -53,14 +53,14 @@ void slice_resize(struct slice *s, int l);
 int slice_compare(struct slice a, struct slice b);
 
 /* Append `data` to the `dest` slice.  */
-int slice_write(struct slice *dest, byte *data, int sz);
+int slice_write(struct slice *dest, byte *data, size_t sz);
 
 /* Append `add` to `dest. */
 void slice_append(struct slice *dest, struct slice add);
 
 /* Like slice_write, but suitable for passing to reftable_new_writer
  */
-int slice_write_void(void *b, byte *data, int sz);
+int slice_write_void(void *b, byte *data, size_t sz);
 
 /* Find the longest shared prefix size of `a` and `b` */
 int common_prefix_size(struct slice a, struct slice b);
